Fixed issues with RJB usage

+++ addon/buildr/scalabison.rb
@@ -6,9 +6,14 @@ module Buildr
   module ScalaBison
     SCALA_BISON = 'edu.uwm.cs:scalabison:jar:0.78'
     
+    Java.classpath << Buildr::Scala::Scalac.dependencies
+    Java.classpath << SCALA_BISON
+    
     def scalabison(*dirs)
       target = _(:target, :generated, :scalabison)
       task = file target
+      
+      Java.load
       
       dirs.each do |dir|
         task.enhance [dir]
@@ -34,18 +39,15 @@ module Buildr
             Dir.chdir dname
             
             trace "Running bison against #{fname}"
-            system 'bison', '-v', '--no-parser', fname
+            unless system 'bison', '-v', '--no-parser', fname
+              fail 'Bison failed to generate LALR tables. See errors above.'
+            end
             
             FileUtils.cp fname, dname
-            
-            Java.classpath << Buildr::Scala::Scalac.dependencies
-            Java.classpath << artifact(SCALA_BISON)
             
             args = ['-v', cname + '.y']
             
             trace "Running ScalaBison against #{cname + '.y'}"
-            
-            Java.load
             Java.edu.uwm.cs.cool.meta.parser.RunGenerator.main(args.to_java(Java.java.lang.String))
             
             FileUtils.rm cname + '.y'
